JAMES-2890 Split JMAP RFC-8621 Serializer
Rationals:

  - Many git conflicts in that class
  - This lengthy class is hard to read with almost no structure
  - Too much interleaving implicits...
  - Some static JSON deserialization requires a val serializer

 This is doable as we perform several parsing passes:

  - first the request structure
  - then back-reference substitutions
  - then method argument resolution
  - the */set creation and patches validation

These stages are independant from each other, and Methods are groupable by mostly distinct sets,
allowing a better modularisation.
