+++ server/protocols/jmap-rfc-8621/src/test/scala/org/apache/james/jmap/json/ResponseObjectSerializationTest.scala
@@ -21,16 +21,10 @@ package org.apache.james.jmap.json
 
 import eu.timepit.refined.auto._
 import org.apache.james.jmap.json.Fixture._
-import org.apache.james.jmap.json.ResponseObjectSerializationTest.SERIALIZER
 import org.apache.james.jmap.model.ResponseObject
-import org.apache.james.mailbox.model.TestId
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import play.api.libs.json._
-
-object ResponseObjectSerializationTest {
-  private val SERIALIZER: Serializer = new Serializer(new TestId.Factory)
-}
 
 class ResponseObjectSerializationTest extends AnyWordSpec with Matchers {
   "Deserialize ResponseObject" should {
@@ -39,7 +33,7 @@ class ResponseObjectSerializationTest extends AnyWordSpec with Matchers {
         sessionState = "75128aab4b1b",
         methodResponses = Seq(invocation1))
 
-      SERIALIZER.deserializeResponseObject(
+      ResponseSerializer.deserializeResponseObject(
         """
           |{
           |  "methodResponses": [
@@ -58,7 +52,7 @@ class ResponseObjectSerializationTest extends AnyWordSpec with Matchers {
         sessionState = "75128aab4b1b",
         methodResponses = Seq(invocation1, invocation2))
 
-      SERIALIZER.deserializeResponseObject(
+      ResponseSerializer.deserializeResponseObject(
         """
           |{
           |  "sessionState": "75128aab4b1b",
@@ -96,7 +90,7 @@ class ResponseObjectSerializationTest extends AnyWordSpec with Matchers {
           |}
           |""".stripMargin))
 
-      SERIALIZER.serialize(responseObject) should be(Json.parse(expectedJson))
+      ResponseSerializer.serialize(responseObject) should be(Json.parse(expectedJson))
     }
   }
 }
