+++ server/protocols/jmap-rfc-8621/src/test/scala/org/apache/james/jmap/json/SessionSerializationTest.scala
@@ -29,7 +29,6 @@ import org.apache.james.jmap.model.CoreCapabilityProperties.CollationAlgorithm
 import org.apache.james.jmap.model.MailCapability.EmailQuerySortOption
 import org.apache.james.jmap.model.State.State
 import org.apache.james.jmap.model._
-import org.apache.james.mailbox.model.TestId
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AnyWordSpec
 import play.api.libs.json.Json
@@ -208,7 +207,7 @@ class SessionSerializationTest extends AnyWordSpec with Matchers {
           |  "eventSourceUrl": "http://james.org",
           |  "state": "fda9342jcm"
           |}""".stripMargin)
-      new Serializer(new TestId.Factory).serialize(SESSION) should equal(json)
+      ResponseSerializer.serialize(SESSION) should equal(json)
     }
   }
 }
